Move Config object into separate file
